requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-crispy-forms==1.14.0
django-import-export==2.7.1
django-rosetta==0.9.8
django-tinymce==3.4.0
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
polib==1.1.1
PyYAML==6.0
requests==2.27.1
sqlparse==0.4.2
tablib==3.2.0
urllib3==1.26.8
whitenoise==6.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
